[WIP] [ci] run dash-licenses as part of GH workflow
Add a "3PP FOSS license check" GitHub workflow, that uses
the Eclipse Foundation dash-licences tool to validate the
licenses of our npm dependencies.

One can also run the license check locally with command:
"yarn license:check"

I have considered 3 alternatives:

1) add the license check as a step of the build job of the
build workflow. The license check needs to report success
or it will fail the build job and the publish job (when
applicable). If the license check reports success, it's
burried in the multiple steps of its job, not easily
noticed.

One advantage is that we can re-use the build previously
done in the same job, to run license check on, so not too
much overhead.

2) add the license check as a new job in the build workflow.
This works better but requires a rebuild of the repo (yarn
install would suffice, but we do a full build instead of
just installing the dependencies). There is a lot of redundency
in the job vs build and publish - we need to re-install node,
python, checkout the repo.

One advantage of this approach is that the license check has
visibility in the PR Ci checks section and can be made to fail
without impacting the other jobs.

3) add a new workflow for the license check

Like option 1, this will require a full rebuild of the repo.
Like option 2, it permits the license check having visibility
in the PR CI checks section and can fail without impacting
other workflows/jobs.

This option has the additional advantage that it's quite
generic (for Theia-based repos) and could be re-used more
easily, without changing the main workflow file.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>