[plugin] update signature for activeSignatureHelp
This is done to follow a change in vscode.d.ts, to keep Theia's
version aligned.

The discrepancy was found with the vscode-theia-comparator tool.
Other similar changes were not detected by the tool, apparently
because the TypeScript tooling was able to confirm them equivalent.
We modified the tool locally to list these cases and also aligned
them.

Finally, there are a couple of adjustments that were required for
the code to compile, where 3 different versions of an interface
existed and were aligned, now no longer are (theia, Monaco and
the plugin system). This forced us to switch to using Monaco's
version: CompletionContext and SignatureHelpContext.

Fixes #11828

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>

[plugin] modify other similar APIs

They were found by locally hacking vscode-theia-comparator.

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>

[plugin] adapt to modified APIs

Signed-off-by: Marc Dumais <marc.dumais@ericsson.com>